feat: decouple gnotest
